Add files via upload
As my beginner project, created a slot machine for betting. Which take input from user in the form of a deposit, how many lines they want to bet, and how much they want to bet on each line. Then randomly choose the Symbols and print the slot machine on the screen. If the user wins it multiplies the bet with the value associated with that symbol and prints the outcome. The code keeps a record of the user balance. If they ran out of balance it print low balance.